Change data model

+++ iBlue-Service/src/com/iblue/model/ParkingAllocInterface.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface ParkingAllocInterface {
 
-	public int getNearestStreetId(SpotInterface spot);
+	public long getNearestStreetId(SpotInterface spot);
 	public List<? extends StreetAvailabilityInterface> getNearStreetAvailability(SpotInterface spot) ;
 	public StreetAvailabilityInterface parkMeClosest(SpotInterface spot) ;
